requirements.txt
pytest
allure-pytest
pyautogui
opencv-python
# Add other deps like pillow, numpy, etc. if needed
